requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain-huggingface>=0.0.6
huggingface_hub>=0.24.0
python-dotenv==1.0.1
pydantic>=2.0.0
openai>=1.0.0
flask==3.1.0
requests==2.32.3
google-api-python-client>=2.97.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
dateparser>=1.1.8 